backend/pyproject.toml
[project]
name = "projet-restau-api"
version = "0.1.0"
dependencies = [
  "fastapi>=0.110",
  "pydantic>=2.6",
  "uvicorn[standard]",
  "pydantic-settings>=2.2",
  "sqlalchemy>=2.0",
  "psycopg[binary]>=3.1",
  "alembic>=1.13",
  "python-multipart>=0.0.9",
  "pandas>=2.2",
  "passlib==1.7.4",
  "bcrypt==4.0.1",
  "openpyxl>=3.1",
  "pytest>=8.0",
  "email-validator",
  "python-jose[cryptography]",
  "httpx>=0.27",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
